Main svc public

+++ main-service/src/main/java/ru/practicum/repository/CompilationRepository.java
@@ -3,5 +3,5 @@ package ru.practicum.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.practicum.entity.Compilation;
 
-public interface CompilationRepository extends JpaRepository<Compilation, Integer> {
+public interface CompilationRepository extends JpaRepository<Compilation, Long> {
 }
